pyproject.toml
[tool.poetry]
name = "Tanks"
version = "0.1.0"
description = ""
authors = ["XetPy1030 <xetpy1030@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.2"}
click = "^8.1.7"
uvicorn = "^0.32.0"
python-socketio = "^5.11.4"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
pytest-anyio = "^0.0.0"
trio = "^0.27.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
